Don't match on opaque types in candidates/2
There was a bug when using sets, where there could be many representations of an
empty set and relaying on matching, to find out if a set is empty caused
generation of fewer candidates. With mapsets implementation this is no longer a
case, but it's still a bug. Checking set's size instead.